PWN-9472 -  Dynamic percent diff

+++ core/src/main/java/org/p2p/core/token/TokenMetadataExtension.kt
@@ -8,7 +8,7 @@ data class TokenMetadataExtension(
     val ruleOfProcessingTokenPriceWs: String? = null,
     val isPositionOnWs: Boolean? = null,
     val isTokenCellVisibleOnWs: Boolean? = null,
-    val percentDifferenceToShowByPriceOnWs: Int? = null,
+    val percentDifferenceToShowByPriceOnWs: Double? = null,
     val calculationOfFinalBalanceOnWs: Boolean? = null,
     val ruleOfFractionalPartOnWs: String? = null,
     val canBeHidden: Boolean? = null
